reorganize tests + adding new tests
